Use C# 7.3 enum constraint. Added disposable extensions. (#4)

+++ Centare.Extensions/TupleExtensions.cs
@@ -21,28 +21,28 @@ namespace Centare.Extensions
             => (result: (ushort)(numerator / denominator), remainder: (ushort)(numerator % denominator));
 
         public static (int result, int remainder) DividedBy(this int numerator, int denominator)
-            => (result: (int)(numerator / denominator), remainder: (int)(numerator % denominator));
+            => (result: numerator / denominator, remainder: numerator % denominator);
 
         public static (uint result, uint remainder) DividedBy(this uint numerator, uint denominator)
-            => (result: (uint)(numerator / denominator), remainder: (uint)(numerator % denominator));
+            => (result: numerator / denominator, remainder: numerator % denominator);
 
         public static (long result, long remainder) DividedBy(this long numerator, long denominator)
-            => (result: (long)(numerator / denominator), remainder: (long)(numerator % denominator));
+            => (result: numerator / denominator, remainder: numerator % denominator);
 
         public static (ulong result, ulong remainder) DividedBy(this ulong numerator, ulong denominator)
-            => (result: (ulong)(numerator / denominator), remainder: (ulong)(numerator % denominator));
+            => (result: numerator / denominator, remainder: numerator % denominator);
 
         public static (float result, float remainder) DividedBy(this float numerator, float denominator)
-            => (result: (float)(numerator / denominator), remainder: (float)(numerator % denominator));
+            => (result: numerator / denominator, remainder: numerator % denominator);
 
         public static (double result, double remainder) DividedBy(this double numerator, double denominator)
-            => (result: (double)(numerator / denominator), remainder: (double)(numerator % denominator));
+            => (result: numerator / denominator, remainder: numerator % denominator);
 
         public static (decimal result, decimal remainder) DividedBy(this decimal numerator, decimal denominator)
-            => (result: (decimal)(numerator / denominator), remainder: (decimal)(numerator % denominator));
+            => (result: numerator / denominator, remainder: numerator % denominator);
 
         public static async Task TryAwaitOrLogAsync(
-            this(Task task, ILogger logger) t)
+            this (Task task, ILogger logger) t)
         {
             try
             {
